+++ src/org/jmol/modelset/HBond.java
@@ -29,9 +29,9 @@ import org.jmol.util.Logger;
 
 public class HBond extends Bond {
 
-  float energy;
+  double energy;
   
-  public HBond(Atom atom1, Atom atom2, int order, short mad, short colix, float energy) {
+  public HBond(Atom atom1, Atom atom2, int order, short mad, short colix, double energy) {
     super(atom1, atom2, order, mad, colix);
     this.energy = energy;
     if (Logger.debugging)
@@ -39,7 +39,7 @@ public class HBond extends Bond {
   }
   
   @Override
-  public float getEnergy() {
+  public double getEnergy() {
     return energy;
   }
